working crud, added authorization

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -14,17 +15,29 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
-  })
+    text: req.body.text,
+    user: req.user.id
+  });
   res.status(200).json(goal);
 });
 
 const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
-
   if (!goal) {
-    res.status(400)
-    throw new Error('Goal not found.')
+    res.status(400);
+    throw new Error('Goal not found.');
+  }
+
+  const user = await User.findById(req.user.id);
+  if(!user) {
+    res.status(401);
+    throw new Error('User not found.');
+  }
+
+  // make sure goal user is the logged user
+  if(goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not autorized.');
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true});
@@ -34,10 +47,21 @@ const updateGoal = asyncHandler(async (req, res) => {
 
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
+  if(!goal) {
+    res.status(400);
+    throw new Error('Goal not found.');
+  }
 
-  if (!goal) {
-    res.status(400)
-    throw new Error('Goal not found.')
+  const user = await User.findById(req.user.id);
+  if(!user) {
+    res.status(401);
+    throw new Error('User not found.');
+  }
+
+  // make sure goal user is the logged user
+  if(goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not autorized.');
   }
 
   await goal.deleteOne({ _id: req.params.id });
